Move other related parts in common pallet
